mybatis : 전박적인 작업 완료

+++ sample-springboot-mybatis/src/main/java/io/simpolor/mybatis/mapper/StudentMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Mapper
 public interface StudentMapper {
 
-	int selectStudentTotalCountMapper();
+	long selectStudentTotalCountMapper();
 
 	List<Student> selectStudentListMapper();
